% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{admt.coll}
\alias{admt.coll}
\title{Major admit to college}
\usage{
admt.coll(pidlist, termid = "All")
}
\arguments{
\item{termid}{default is All to pull all available terms from SISFUll or a character vector to specify term to pull the major from}
}
\value{
A data frame with enrolled Pids and major by term and admit to college status
}
\description{
admt.coll is a function to pull undergraduates all majors by term and major college from SISFull, and indicate whether major is admitted to college or not
}
