% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{cohort}
\alias{cohort}
\title{cohort is a function to pull the entering cohorts for defined population. Fall cohort including summer starters}
\usage{
cohort(ds = "SISFull", pidlist)
}
\arguments{
\item{pidlist}{a character vector of Pids interested}
}
\value{
a dataframe with Pid, cohort, Entry Term, AidYr Info
}
\description{
cohort is a function to pull the entering cohorts for defined population. Fall cohort including summer starters
}
