% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{FA_pull}
\alias{FA_pull}
\title{FA_pull is a function to get Pell, FAFSA 1st Gen, and other FA related Info by Pid}
\usage{
FA_pull(pidlist)
}
\arguments{
\item{pidlist}{a character vectors of Pids}
}
\value{
a data frame with Pid, entry cohort, and FA info
}
\description{
FA_pull is a function to get Pell, FAFSA 1st Gen, and other FA related Info by Pid
}
