% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{firstATL_pull}
\alias{firstATL_pull}
\title{firstATL_pull is a function to pull First ATL Course and Grade}
\usage{
firstATL_pull(ds = "SISFull", pidlist)
}
\arguments{
\item{ds}{A character string to indicate the SIS source : SISFull or SISInfo}

\item{pidlist}{A character vector of Pids}
}
\value{
A data frame with Pids, Term_Seq_Id when first ATL was taken, 1st ATL and Grade
}
\description{
firstATL_pull is a function to pull First ATL Course and Grade
}
