% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{firstgen_add}
\alias{firstgen_add}
\title{firstgen_add is a function to add students with the first generation status to the dataframe provided}
\usage{
firstgen_add(ds = "SISFull", maindat)
}
\arguments{
\item{ds}{A character string to indicate the SIS source : SISFull or SISInfo}

\item{maindat}{A dataframe with identifer as either Pid or PID}
}
\value{
this function add another column FGEN to the original dataframe to indicate first gen status
}
\description{
firstgen_add is a function to add students with the first generation status to the dataframe provided
}
\examples{
\dontrun{firstgen_add(ds='SISFull', maindat=PAG)}

}
