% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{firstgen_pull}
\alias{firstgen_pull}
\title{Admission First Gen}
\usage{
firstgen_pull(ds = "SISFull", pidlist)
}
\arguments{
\item{ds}{A character string to indicate the SIS source : SISFull or SISInfo}

\item{pidlist}{a character vector to input the population interested identifying by Pid}
}
\value{
this function returns a dataframe with only firstGen students
}
\description{
firstgen_pull is a function to return student records with the admission first generation status from SISAPRS.first_gen is available after entering 2008 for UN and AT
}
