% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{firsttermgpa_pull}
\alias{firsttermgpa_pull}
\title{firsttermgpa_pull is a function to pull First Term GPA by student level}
\usage{
firsttermgpa_pull(ds = "SISFull", pidlist)
}
\arguments{
\item{ds}{A character string to indicate the SIS source : SISFull or SISInfo}

\item{pidlist}{A character vector of Pids}
}
\value{
A data frame with Pids, student_level,first_term_seq_id, termGPA, firstterm GPA credits
}
\description{
firsttermgpa_pull is a function to pull First Term GPA by student level
}
