% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{gndr_race_pull}
\alias{gndr_race_pull}
\title{Gender and Race Ethnicity}
\usage{
gndr_race_pull(ds = "SISFull", pidlist, ex = "QRTRTERM")
}
\arguments{
\item{ds}{A character string to indicate the SIS source : SISFull or SISInfo or SISFrzn}

\item{pidlist}{A character vector of Pids}

\item{ex}{a character string to indicate which flavor of the SISFrzn extract to use,value can only be within QRTRTERM,FIRSTDAY,ENDTERM}
}
\value{
A data frame with Pids, Gender, Ethnicity and Ctzn Code. If ds is SISFrzn, the status is by term
}
\description{
gndr_race_pull is a function to provide gender, IPEDS ethnicity for the Pids
}
