% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{honor_pull}
\alias{honor_pull}
\title{Honor and Academic Scholar}
\usage{
honor_pull(ds = "SISFull", pidlist)
}
\arguments{
\item{ds}{A character string to indicate the SIS source : SISFull or SISInfo}

\item{pidlist}{a character vector to input the population interested identifying by Pid}
}
\value{
this function returns a dataframe with combined honor or academic scholar status by term
}
\description{
honor_pull is a function to get student with the honor or academic scholar combined status by term
}
