% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{indicator_pull}
\alias{indicator_pull}
\title{indicator_pull is a function to pull indicator and atrisk for undergrads and AT from SISFull}
\usage{
indicator_pull(pidlist)
}
\arguments{
\item{pidlist}{A character vector of Pids}
}
\value{
A data frame with Pids, indicator,atrisk,FirsttermGPA, student_level, pred_GPA, 1stATL
}
\description{
indicator_pull is a function to pull indicator and atrisk for undergrads and AT from SISFull
}
