% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{Preliminary.PFS2}
\alias{Preliminary.PFS2}
\title{Preliminary.PFS2 is a function to get the 1st Spring and 2nd Fall persistence for undergraduate fall entering cohort including summer starters from SISFrzn extracts}
\usage{
Preliminary.PFS2(cohortex = "QRTRTERM", subenrlex = "FIRSTDAY", cohortterm,
  output = "aggregate")
}
\arguments{
\item{cohortex}{a character string to indicate the fall entering cohort was built from which flavor of the SISFrzn extract,value can only be within QRTRTERM,FIRSTDAY,ENDTERM}

\item{subenrlex}{a character string to indicate the subsequent enroll and Award population was pulled from which flavor of the SISFrzn extract,value can only be within QRTRTERM,FIRSTDAY,ENDTERM}

\item{cohortterm}{a numeric vector to indicate the fall entering cohort 1st Fall term,e.g 1144}

\item{output}{indicate the output format, aggregate or list}
}
\value{
aggregate output returns the persistence rate by entering cohort and lvl entry status. list returns the detail data frame
}
\description{
Preliminary.PFS2 is a function to get the 1st Spring and 2nd Fall persistence for undergraduate fall entering cohort including summer starters from SISFrzn extracts
}
