% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{term.enroll}
\alias{term.enroll}
\title{Term Enrollment Student List}
\usage{
term.enroll(ds = "SISFrzn", ex = "QRTRTERM", termid)
}
\arguments{
\item{ds}{A character string to indicate the SIS source  SISFrzn or SISFull or SISInfo}

\item{ex}{a character string to indicate which flavor of the SISFrzn extract to use,value can only be within QRTRTERM,FIRSTDAY,ENDTERM}

\item{termid}{a numeric vector to specify which term or terms to pull the data from}
}
\value{
A data frame with enrolled Pids and term related info
}
\description{
term.enroll is a function to get enrollment unit record data including official Frzn enrollment
}
