% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{cohort}
\alias{cohort}
\title{Entering cohort Info}
\usage{
cohort(ds = "SISFull", ex = "QRTRTERM", pidlist)
}
\arguments{
\item{ds}{A character string to indicate the SIS source  SISFrzn or SISFull or SISInfo}

\item{ex}{a character string to indicate which flavor of the SISFrzn extract to use,value can only be within QRTRTERM,FIRSTDAY,ENDTERM}

\item{pidlist}{a character vector of Pids interested}
}
\value{
a dataframe with Pid, student_level, cohort, Entry Term, AidYr Info
}
\description{
cohort is a function to get the entering cohorts for defined population including all levels. Fall cohort including summer starters
}
