% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{gndr_race_pull}
\alias{gndr_race_pull}
\title{gndr_race_pull is a function to provide gender, IPEDS ethnicity for the Pids provided}
\usage{
gndr_race_pull(ds = "SISFull", pidlist)
}
\arguments{
\item{ds}{A character string to indicate the SIS source : SISFull or SISInfo}

\item{pidlist}{A character vector of Pids}
}
\value{
A data frame with Pids, Gender, Ethnicity and Ctzn Code
}
\description{
gndr_race_pull is a function to provide gender, IPEDS ethnicity for the Pids provided
}
