% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{honor_pull}
\alias{honor_pull}
\title{honor_pull is a function to return students with the honor or academic scholar status by term}
\usage{
honor_pull(ds = "SISFull", pidlist)
}
\arguments{
\item{ds}{A character string to indicate the SIS source : SISFull or SISInfo}

\item{pidlist}{a character vector to input the population interested identifying by Pid}
}
\value{
this function returns a dataframe with only honor or academic scholar students by term
}
\description{
honor_pull is a function to return students with the honor or academic scholar status by term
}
\examples{
\dontrun{honor_pull(ds='SISFull', pidlist=PAG$PID)}

}
