% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{mjrclass_pull}
\alias{mjrclass_pull}
\title{mjrclass_pull is a function to return student term info on class code and the primary major}
\usage{
mjrclass_pull(ds = "SISFull", pidlist)
}
\arguments{
\item{ds}{A character string to indicate the SIS source : SISFull or SISInfo}

\item{pidlist}{a character vector to input the population interested identifying by Pid}
}
\value{
this function returns a dataframe with only honor or academic scholar students by term
}
\description{
mjrclass_pull is a function to return student term info on class code and the primary major
}
\examples{
\dontrun{mjrclass_pull(ds='SISFull', pidlist=PAG$PID)}

}
