% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulldata.R
\name{msudatacon}
\alias{msudatacon}
\title{msudatacon is a function to allow connect to MSUDATA using JDBC}
\usage{
msudatacon(userid)
}
\arguments{
\item{userid}{a character value to indiate userid for MSUDATA}
}
\value{
MSUDATA JDBC connection
}
\description{
msudatacon is a function to allow connect to MSUDATA using JDBC
}
